updated daemon long/int types
Former-commit-id: 38ba8627a2ad80c162f488548d821cecab41b562

+++ src/main/java/monero/daemon/model/MoneroBlock.java
@@ -239,7 +239,7 @@ public class MoneroBlock extends MoneroBlockHeader {
   }
   
   @Override
-  public MoneroBlock setNonce(Long nonce) {
+  public MoneroBlock setNonce(Integer nonce) {
     super.setNonce(nonce);
     return this;
   }
